Fix: error using the visit mothod.
Call the method visit in a wizard with some pages, and then press the
Next and Prev buttons a couple of times to reproduce the error.

Problem: the method visit must remove from the containers
(_cancelStatuses and the others) the information of the current step,
but it is performing TWO "pop" operations for every container.
This removes too many things and raises NPE because the containers do
not have the correct number of steps.